server/src/services/ocr-worker/requirements.txt
opencv-python
pytesseract
Pillow
redis
numpy
python-dotenv
flask
opencv-python
requests
